Añadidas más integraciones con el framework gob.cl
se usan más estilos para el frontend provenientes del framework gob.cl, además se teermina la logica CRUD para el fondo

+++ frontend/src/routes/Fondos/DeleteFondo.jsx
@@ -1,25 +1,49 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { deleteFondo } from "../../services/fondo.service.js";
+import Swal from 'sweetalert2';
+
 
 const DeleteFondo = ({ id }) => {
   const [isDeleting, setIsDeleting] = useState(false);
   const navigate = useNavigate();
   const handleDelete = async () => {
-    setIsDeleting(true);
-    try {
-      await deleteFondo(id);
-      // Si la eliminación es exitosa, redirigir o actualizar la lista de fondos
-      navigate("/fondos"); // Asumiendo que "/fondos" es la ruta donde se lista los fondos
-    } catch (error) {
-      console.error("Error al eliminar el fondo:", error);
-      // Manejar el error (mostrar mensaje al usuario, etc.)
-    } finally {
-      setIsDeleting(false);
-    }
-  };
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: "¡No podrás revertir esta acción!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sí, eliminarlo!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        setIsDeleting(true);
+        deleteFondo(id)
+          .then(() => {
+            Swal.fire(
+              '¡Eliminado!',
+              'El fondo ha sido eliminado.',
+              'success'
+            );
+            navigate("/fondos");
+          })
+          .catch((error) => {
+            console.error("Error al eliminar el fondo:", error);
+            Swal.fire(
+              'Error',
+              'No se pudo eliminar el fondo.',
+              'error'
+            );
+          })
+          .finally(() => {
+            setIsDeleting(false);
+          });
+      }
+    });
+  };  
   return (
-    <button class="btn btn-pill-secondary" onClick={handleDelete} disabled={isDeleting}>
+    <button className="btn btn-pill-secondary" onClick={handleDelete} disabled={isDeleting}>
       {isDeleting ? "Eliminando..." : "Eliminar Fondo"}
     </button>
   );
